Register interface & sharing components between login and register screens
