+++ projects/desktop/src/components/LoginScreen/index.tsx
@@ -1,6 +1,8 @@
 import React from 'react'
-import {BigTitle, LoginLanguageChooser, Logo} from './styles'
-import { Container, LeftSide, RightSide } from './parts'
+import {Box, LoginLanguageChooser, Logo} from './styles'
+import { SingleFormContainer } from '../common/Containers/SingleFormContainer'
+import { BigTitle } from '../common/SingleForm'
+import { LeftSide, RightSide } from './parts'
 import { Form as FinalForm } from 'react-final-form'
 import { Form } from './form'
 import {useTranslation, UseTranslationResponse} from 'react-i18next'
@@ -14,21 +16,23 @@ export const LoginScreen = (): JSX.Element => {
     }
     
     return (
-        <Container>
-            <LeftSide>
-                <Logo src={logo} />
-                <LoginLanguageChooser />
-                <BigTitle dangerouslySetInnerHTML={getTitleContent()} />
+        <SingleFormContainer>
+            <Box>
+                <LeftSide>
+                    <Logo src={logo} alt='Bitsky' />
+                    <LoginLanguageChooser />
+                    <BigTitle dangerouslySetInnerHTML={getTitleContent()} />
 
-                <FinalForm
-                    onSubmit={values => console.log(values)}
-                    initialValues={{ remember: false }}
-                    render={({handleSubmit}) => <Form handleSubmit={handleSubmit} />}
-                />
-            </LeftSide>
-            <RightSide>
-                <br/>
-            </RightSide>
-        </Container>
+                    <FinalForm
+                        onSubmit={values => console.log(values)}
+                        initialValues={{ remember: false }}
+                        render={({handleSubmit}) => <Form handleSubmit={handleSubmit} />}
+                    />
+                </LeftSide>
+                <RightSide>
+                    <br/>
+                </RightSide>
+            </Box>
+        </SingleFormContainer>
     )
 }
