--- projects/desktop/src/components/LoginScreen/parts/Container.tsx
@@ -1,14 +1,0 @@
-import React from 'react'
-import { Background, WhiteOverlay, Box, LoginContainer } from '../styles'
-
-export const Container = ({children}: IPropsChildren): JSX.Element => (
-    <Background>
-        <WhiteOverlay>
-            <LoginContainer maxWidth='lg'>
-                <Box>
-                    {children}
-                </Box>
-            </LoginContainer>
-        </WhiteOverlay>
-    </Background>
-)
