+++ projects/desktop/src/components/LoginScreen/parts/index.ts
@@ -1,3 +1,2 @@
-export { Container } from './Container'
 export { LeftSide } from './LeftSide'
 export { RightSide } from './RightSide'
