+++ projects/desktop/src/components/RegisterScreen/form/index.ts
@@ -1,0 +1,3 @@
+export { Form } from './Form'
+export { FirstnameField } from './FirstnameField'
+export { LastnameField } from './LastnameField'
