+++ projects/desktop/src/components/common/Background/WhiteOverlay.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const WhiteOverlay = styled.div`
+    background-color: rgba(255, 255, 255, 0.6);
+`
