+++ projects/desktop/src/components/common/Background/index.ts
@@ -1,0 +1,2 @@
+export { Background } from './Background'
+export { WhiteOverlay } from './WhiteOverlay'
