+++ projects/desktop/src/components/common/Containers/FullHeightContainer.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+import {Container as MaterialContainer} from '@material-ui/core'
+
+export const FullHeightContainer = styled(MaterialContainer)`
+    min-height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+`
