+++ projects/desktop/src/components/common/SingleForm/input/InputContainer.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const InputContainer = styled.div`
+    display: flex;
+    width: 100%;
+    margin-top: 16px;
+`
