+++ projects/desktop/src/components/common/SingleForm/input/InputError.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+import {colors} from '../../../../constants'
+
+export const InputError = styled.span`
+    font-family: 'Montserrat', sans-serif;
+    position: absolute;
+    bottom: -22px;
+    left: 0;
+    color: ${colors.error};
+`
