+++ projects/desktop/src/components/common/SingleForm/input/InputField.tsx
@@ -1,0 +1,28 @@
+import styled from 'styled-components'
+import {TextField} from '@material-ui/core'
+import {colors} from '../../../../constants'
+
+export const InputField = styled(TextField)`
+    width: calc(100% - 2px);
+    font-family: sans-serif;
+
+    fieldset {
+        border-left: none;
+        border-radius: 0 8px 8px 0;
+        border-color: ${colors.lightGrey};
+    }
+
+    label {
+        color: ${colors.grey};
+    }
+
+    label.Mui-focused {
+        color: ${colors.grey} !important;
+    }
+
+    div:hover, .Mui-focused {
+        fieldset {
+            border: 1px solid ${colors.lightGrey} !important;
+        }
+    }
+`
