+++ projects/desktop/src/components/common/SingleForm/input/InputsRow.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+
+export const InputsRow = styled.div`
+  display: flex;
+  
+  > * {
+    margin-right: 30px;
+    min-width: 300px;
+  }
+  
+  > :last-child {
+    margin-right: 0;
+  }
+`
